+++ crates/wasi-otel/src/guest/convert.rs
@@ -1,13 +1,50 @@
 //! # Convert
 //!
-//! Convert OpenTelemetry types in `wasi-otel` types.
+//! Convert between OpenTelemetry types and `wasi-otel` types.
 
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use opentelemetry::{Array, InstrumentationScope, Key, KeyValue, Value};
+use opentelemetry_sdk::Resource;
 
 use crate::guest::generated::omnia::otel::types as wasi;
 use crate::guest::generated::wasi::clocks::wall_clock::Datetime;
+
+impl From<wasi::Resource> for Resource {
+    fn from(value: wasi::Resource) -> Self {
+        let attrs = value.attributes.into_iter().map(Into::into).collect::<Vec<_>>();
+        let builder = Self::builder();
+
+        if let Some(schema_url) = value.schema_url {
+            builder.with_schema_url(attrs, schema_url).build()
+        } else {
+            builder.with_attributes(attrs).build()
+        }
+    }
+}
+
+impl From<wasi::KeyValue> for KeyValue {
+    fn from(value: wasi::KeyValue) -> Self {
+        Self::new(value.key, value.value)
+    }
+}
+
+impl From<wasi::Value> for Value {
+    fn from(value: wasi::Value) -> Self {
+        match value {
+            wasi::Value::Bool(v) => Self::Bool(v),
+            wasi::Value::S64(v) => Self::I64(v),
+            wasi::Value::F64(v) => Self::F64(v),
+            wasi::Value::String(v) => Self::String(v.into()),
+            wasi::Value::BoolArray(items) => Self::Array(Array::Bool(items)),
+            wasi::Value::S64Array(items) => Self::Array(Array::I64(items)),
+            wasi::Value::F64Array(items) => Self::Array(Array::F64(items)),
+            wasi::Value::StringArray(items) => {
+                Self::Array(Array::String(items.into_iter().map(Into::into).collect()))
+            }
+        }
+    }
+}
 
 impl From<Value> for wasi::Value {
     fn from(value: Value) -> Self {